# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 班导师管理系统 API
  version: 0.1.0
  description: |
    面向前端的后端接口定义。实现以此为准，运行时 springdoc 导出的 openapi.json 仅代表当前已实现接口。
servers:
  - url: http://localhost:8080
...
paths:
...
  /plans:
    get:
      tags: [plans]
      summary: 计划列表（个人/班级）
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: term
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagePlan"
        "401":
          $ref: "#/components/responses/Error401"
    post:
      tags: [plans]
      summary: 创建计划
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Error401"
        "422":
          $ref: "#/components/responses/Error422"
  /plans/{planId}:
    get:
      tags: [plans]
      summary: 计划详情（含条目）
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanDetail"
        "401":
          $ref: "#/components/responses/Error401"
        "404":
          $ref: "#/components/responses/Error404"
  /plans/{planId}/items:
    post:
      tags: [plans]
      summary: 新增计划条目
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanItem"
        "401":
          $ref: "#/components/responses/Error401"
        "404":
          $ref: "#/components/responses/Error404"
        "422":
          $ref: "#/components/responses/Error422"
  /plans/{planId}/items/{itemId}:
    patch:
      tags: [plans]
      summary: 更新计划条目
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: itemId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemUpdateRequest"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Error401"
        "404":
          $ref: "#/components/responses/Error404"
        "422":
          $ref: "#/components/responses/Error422"
    delete:
      tags: [plans]
      summary: 删除计划条目
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: itemId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Error401"
        "404":
          $ref: "#/components/responses/Error404"
  /plan-items/{itemId}/progress:
    get:
      tags: [plans]
      summary: 条目进度历史
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanItemProgress"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
    post:
      tags: [plans]
      summary: 追加条目进度记录
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemProgressCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanItemProgress"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "422":
          $ref: "#/components/responses/Error422"
...
  /reports/plan-completion:
    get:
      tags: [reports]
      summary: 计划完成率报表
      security:
        - bearerAuth: []
      parameters:
        - name: term
          in: query
          required: true
          schema: { type: string }
        - name: collegeId
          in: query
          required: false
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportPlanCompletion"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "422":
          $ref: "#/components/responses/Error422"
  /reports/activity-stats:
    get:
      tags: [reports]
      summary: 活动报名统计
      security:
        - bearerAuth: []
      parameters:
        - name: term
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportActivityStats"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "422":
          $ref: "#/components/responses/Error422"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
  responses:
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            invalid:
              value: { code: AUTH_INVALID_CREDENTIALS, message: invalid credentials }
    Error403:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            forbidden:
              value: { code: AUTH_FORBIDDEN, message: forbidden }
    Error404:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            notFound:
              value: { code: NOT_FOUND, message: not found }
...
    Error422:
      description: Unprocessable Entity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            validation:
              value: { code: VALIDATION_ERROR, message: validation error }
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: field 'x' is required
        requestId:
          type: string
        details:
          type: object
...
    PlanItemStatus:
      type: string
      enum: [todo, doing, done]
    PlanOwnerType:
      type: string
      enum: [CLASS, USER]
    Plan:
      type: object
      required: [id, ownerType, term, title, createdAt]
      properties:
        id: { type: integer, format: int64 }
        ownerType: { $ref: "#/components/schemas/PlanOwnerType" }
        ownerUserId: { type: integer, format: int64, nullable: true }
        ownerClassId: { type: integer, format: int64, nullable: true }
        term: { type: string }
        title: { type: string }
        createdAt: { type: string, format: date-time }
    PlanCreateRequest:
      oneOf:
        - type: object
          required: [ownerType, term, title]
          properties:
            ownerType: { type: string, enum: [USER] }
            term: { type: string }
            title: { type: string }
        - type: object
          required: [ownerType, ownerClassId, term, title]
          properties:
            ownerType: { type: string, enum: [CLASS] }
            ownerClassId: { type: integer, format: int64 }
            term: { type: string }
            title: { type: string }
    PlanItem:
      type: object
      required: [id, planId, title, status, sortOrder, createdAt]
      properties:
        id: { type: integer, format: int64 }
        planId: { type: integer, format: int64 }
        title: { type: string }
        status: { $ref: "#/components/schemas/PlanItemStatus" }
        sortOrder: { type: integer }
        dueDate: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PlanItemCreateRequest:
      type: object
      required: [title]
      properties:
        title: { type: string }
        dueDate: { type: string, format: date, nullable: true }
    PlanItemUpdateRequest:
      type: object
      properties:
        title: { type: string }
        status: { $ref: "#/components/schemas/PlanItemStatus" }
        sortOrder: { type: integer }
        dueDate: { type: string, format: date, nullable: true }
    PlanProgress:
      type: object
      description: 实时计算的计划完成率（以 plan_item.status 为准）
      required: [planId, doneCount, totalCount, completionRate, calculatedAt]
      properties:
        planId: { type: integer, format: int64 }
        doneCount: { type: integer }
        totalCount: { type: integer }
        completionRate: { type: number, format: float }
        calculatedAt: { type: string, format: date-time }
    PlanDetail:
      allOf:
        - $ref: "#/components/schemas/Plan"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/PlanItem" }
            progress:
              $ref: "#/components/schemas/PlanProgress"
    PlanItemProgress:
      type: object
      description: 条目进度历史记录（append-only）
      required: [id, planItemId, percent, createdByUserId, createdAt]
      properties:
        id: { type: integer, format: int64 }
        planItemId: { type: integer, format: int64 }
        percent: { type: integer, minimum: 0, maximum: 100 }
        note: { type: string, nullable: true }
        createdByUserId: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
    PlanItemProgressCreateRequest:
      type: object
      required: [percent]
      properties:
        percent: { type: integer, minimum: 0, maximum: 100 }
        note: { type: string, nullable: true }
...
    ReportPlanCompletion:
      type: object
      required: [term, items]
      properties:
        term: { type: string }
        items:
          type: array
          items:
            type: object
            required: [scope, doneCount, totalCount, completionRate]
            properties:
              scope:
                type: string
                description: 统计范围描述（如“学院:xx”或“班级:xx”）
              doneCount: { type: integer }
              totalCount: { type: integer }
              completionRate: { type: number, format: float }
    ReportActivityStats:
      type: object
      required: [term, items]
      properties:
        term: { type: string }
        items:
          type: array
          items:
            type: object
            required: [activityId, title, approvedCount, appliedCount]
            properties:
              activityId: { type: integer, format: int64 }
              title: { type: string }
              appliedCount: { type: integer }
              approvedCount: { type: integer }
...
    PagePlan:
      type: object
      required: [page, size, total, items]
      properties:
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/Plan" }
...